refactor: exclude default roles by default
