--- src/lib/constants/schemas.ts
@@ -1,7 +1,0 @@
-export const DEFAULT_SYSTEM_SCHEMAS: string[] = [
-  'information_schema',
-  'pg_catalog',
-  'pg_toast_temp_1',
-  'pg_temp_1',
-  'pg_toast',
-]
